ai-test-agent-system/pyproject.toml
[project]
name = "ai-test-agent-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.12",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph-cli[inmem]>=0.4.15",
    "pillow>=12.1.1",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "deepagents>=0.4.12",
    "langchain-pymupdf4llm>=0.5.0",
    "pymupdf4llm>=0.0.17",
    "asyncio>=4.0.0",
    "fastmcp>=3.1.1",
    "langchain-ollama>=1.0.1",
    "openpyxl>=3.1.5",
    "playwright>=1.58.0",
    "python-levenshtein>=0.27.3",
]
